requirements.txt
certifi>=2019.6.16
colorama>=0.4.1
PySocks>=1.7.0
requests>=2.22.0
requests-futures>=1.0.0
stem>=1.8.0 
torrequest>=0.1.0
pandas>=1.0.0
openpyxl<=3.0.10
discord.py>=1.7.3
requests>=2.28.1
line_profiler>=3.5.1
beautifulsoup4>=4.11.1